Move signal functions to separate module, set up sending on stream
